requirements.txt
llama-cpp-python
openai
print-color